Rename files in qtviewer folder
