+++ napari/_qt/qtviewer/qt_controls.py
