+++ napari/_qt/qtviewer/qt_layer_buttons.py
